NLS: Add 6 Multi_key sequences for Ezh
Ezh is a Latin-Script letter belonging to several Uralic, Caucasian,
and West-African languages. It is present on some Finnish keyboards,
but users of many other layouts cannot presently type it. This commit
adds Multi_key sequences for both Capital and lowercase Ezh, as well
as Multi_key + dead_caron sequences for Ezh with a caron, which is
used in Laz and Skolt Sámi.

Part-of: <https://gitlab.freedesktop.org/xorg/lib/libx11/-/merge_requests/221>