nls: add Arabic hamza compose sequences
These sequences are intended for use in the ara(mac-phonetic) and
my(phonetic) layouts. They are based on the following layouts listed in
the CLDR:

- https://github.com/unicode-org/cldr/blob/release-43/keyboards/osx/ar-t-k0-osx-qwerty.xml
- https://github.com/unicode-org/cldr/blob/release-43/keyboards/osx/ms-t-k0-osx.xml

The sequences are listed in the `<transforms>` section, and are
reproduced below:

```
<transforms type="simple">
	<transform from="ء\u{64E}" to="آ"/> <!--  ءَ → آ -->
	<transform from="ء\u{650}" to="إ"/> <!--  ءِ → إ -->
	<transform from="ء " to="ء"/>
	<transform from="ء\u{A0}" to="ء"/>
	<transform from="ء!" to="إ"/>
	<transform from="ء١" to="إ"/>
	<transform from="ءا" to="أ"/>
	<transform from="ءس" to="ئ"/>
	<transform from="ءو" to="ؤ"/>
	<transform from="ءي" to="ئ"/>
	<transform from="ءى" to="ئ"/>
</transforms>
```

We limit ourselves to the sequences that strictly combine a character
and a hamza, and generate that character with a hamza on it, following
the behavior in sequences of other dead keys. Additional sequences,
potentially for other layouts as well, could be added later on as
necessary.

Part-of: <https://gitlab.freedesktop.org/xorg/lib/libx11/-/merge_requests/218>